# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SNM API",
    "description": "API per il progetto \"Social Network for Music\"",
    "version": "1.0.0"
  },
  "host": "localhost:3004",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/playlist/addTrack": {
      "post": {
        "description": "Aggiunta di una nuova traccia alla playlist",
        "parameters": [
          {
            "name": "track_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "playlist_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/deleteTrack": {
      "delete": {
        "description": "Eliminazione di una traccia dalla playlist",
        "parameters": [
          {
            "name": "track_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "playlist_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/createPlaylist": {
      "post": {
        "description": "Creazione di una nuova playlist",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "descrizione": {
                  "example": "any"
                },
                "privata": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/getFollowedPlaylists": {
      "get": {
        "description": "Recupero delle playlist seguite dall'utente",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/getPublicPlaylists": {
      "get": {
        "description": "recupero delle playlists pubbliche",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/getMyPlaylists": {
      "get": {
        "description": "recupero delle playlist dell'utente",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/editPlaylist/{playlist_id}": {
      "patch": {
        "description": "Modifica di una playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "descrizione": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                },
                "privata": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/getPlaylistData/{playlist_id}": {
      "get": {
        "description": "Recupero dei dati di una playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/deletePlaylist/{playlist_id}": {
      "delete": {
        "description": "Eliminazione di una playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/followPlaylist/{playlist_id}": {
      "post": {
        "description": "Permette all'utente di seguire una playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/playlist/unfollowPlaylist/{playlist_id}": {
      "delete": {
        "description": "Permette all'utente di smettere di seguire una playlist",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}